Fixed some of codacy issues
